[fix]: Optimize code
