+++ packages/sqle/src/page/SqlInsights/components/RelatedSqlList/SqlFingerprintColumn.tsx
@@ -29,12 +29,15 @@ const SqlFingerprintColumn: React.FC<SqlFingerprintColumnProps> = ({
       title={t('sqlInsights.relatedSqlList.sqlFingerprintDetail.title')}
       placement="topLeft"
     >
-      <SQLRenderer.Snippet
-        sql={sqlFingerprint}
-        rows={1}
-        tooltip={false}
-        cuttingLength={200}
-      />
+      {/*因为Snippet本身被tooltip包裹，所以这里需要使用div包裹，否则外层的Popover会失效*/}
+      <div>
+        <SQLRenderer.Snippet
+          sql={sqlFingerprint}
+          rows={1}
+          tooltip={false}
+          cuttingLength={200}
+        />
+      </div>
     </Popover>
   );
 };
